Rename ProcessAsyncHelper to AsyncProcessRunner and move it to Utils
